requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
PyPDF2==3.0.1
pdfplumber==0.11.0
Werkzeug==3.0.1
gunicorn==21.2.0
markdown==3.5.1
requests==2.31.0
jsonschema==4.20.0
